Expose the combinator selectors via the visitor.
