+++ src/AngleSharp/AngleSharp.Core.csproj
@@ -591,7 +591,7 @@
     <Compile Include="Css\Dom\Internal\ComplexSelector.cs" />
     <Compile Include="Css\Dom\Internal\CompoundSelector.cs" />
     <Compile Include="Css\Dom\Internal\ListSelector.cs" />
-    <Compile Include="Css\Dom\Internal\SimpleSelector.cs" />
+    <Compile Include="Css\Dom\Internal\NamespaceSelector.cs" />
     <Compile Include="Dom\Internal\DocumentFragment.cs" />
     <Compile Include="Dom\Internal\DomImplementation.cs" />
     <Compile Include="Dom\Internal\EntityReference.cs" />
